build: exclude ui from `test-invoke-pip`
Prior to the folder restructure, the `paths` for `test-invoke-pip` did not include the UI's path `invokeai/frontend/`:

```yaml
    paths:
      - 'pyproject.toml'
      - 'ldm/**'
      - 'invokeai/backend/**'
      - 'invokeai/configs/**'
      - 'invokeai/frontend/dist/**'
```

After the restructure, more code was moved into the `invokeai/frontend/` folder, and `paths` was updated:

```yaml
    paths:
      - 'pyproject.toml'
      - 'invokeai/**'
      - 'invokeai/backend/**'
      - 'invokeai/configs/**'
      - 'invokeai/frontend/web/dist/**'
```

Now, the second path includes the UI. The UI now needs to be excluded, and must be excluded prior to `invokeai/frontend/web/dist/**` being included.

On `test-invoke-pip-skip`, we need to do a bit of logic juggling to invert the folder selection. First, include the web folder, then exclude everying around it and finally exclude the `dist/` folder